+++ Sql/dbo/Tables/Invoices.sql
@@ -2,9 +2,9 @@
 (
 	[invoiceId] INT NOT NULL PRIMARY KEY,
     [userId] INT NOT NULL, 
-    [subtotal] FLOAT NOT NULL, 
-    [tax] FLOAT NOT NULL, 
-    [total] FLOAT NOT NULL, 
+    [subtotal] MONEY NOT NULL, 
+    [tax] MONEY NOT NULL, 
+    [total] MONEY NOT NULL, 
     [datedue] DATE NOT NULL, 
     [datecreated] DATETIME NOT NULL DEFAULT GETDATE() 
 )
